+++ src_plugins/ppx_deriving_eq.ml
@@ -10,7 +10,7 @@ let raise_errorf = Ppx_deriving.raise_errorf
 
 let () =
   Ppx_deriving.register "Eq" (fun ~options ~path type_decls ->
-    let expr_of_type ({ ptype_name = { txt = name }; ptype_loc = loc } as type_) =
+    let expr_of_type ({ ptype_name = { txt = name }; ptype_loc = loc } as type_decl) =
       let argn kind = Printf.sprintf (match kind with `lhs -> "lhs%d" | `rhs -> "rhs%d") in
       let pattn side typs = List.mapi (fun i _ -> pvar (argn side i)) typs in
       let rec exprsn typs =
@@ -47,7 +47,7 @@ let () =
               | _ -> false]
           | { ptyp_desc = Ptyp_constr ({ txt = lid }, args) } ->
             (* ppx_tools#10 *)
-            let fn = Exp.ident (mknoloc (Ppx_deriving.mangle_lid ~prefix:"equal_" lid)) in
+            let fn = Exp.ident (mknoloc (Ppx_deriving.mangle_lid (`Prefix "equal") lid)) in
             (match args with [] -> fn | _ -> app fn (List.map expr_of_typ args))
           | { ptyp_desc = Ptyp_tuple typs } ->
             [%expr fun [%p ptuple (pattn `lhs typs)] [%p ptuple (pattn `rhs typs)] ->
@@ -78,7 +78,7 @@ let () =
                          (Ppx_deriving.string_of_core_type typ)
       in
       let comparator =
-        match type_.ptype_kind, type_.ptype_manifest with
+        match type_decl.ptype_kind, type_decl.ptype_manifest with
         | Ptype_abstract, Some manifest -> expr_of_typ manifest
         | Ptype_variant constrs, _ ->
           let cases =
@@ -102,14 +102,15 @@ let () =
         | Ptype_open, _ ->
           raise_errorf ~loc "Cannot derive Eq for open type"
       in
-      let polymorphize = Ppx_deriving.poly_fun_of_type_decl type_ in
-      [Vb.mk (pvar ("equal_"^name)) (polymorphize comparator)]
+      let polymorphize = Ppx_deriving.poly_fun_of_type_decl type_decl in
+      [Vb.mk (pvar (Ppx_deriving.mangle_type_decl (`Prefix "equal") type_decl))
+             (polymorphize comparator)]
     in
-    let sig_of_type type_ =
-      let typ = Ppx_deriving.core_type_of_type_decl type_ in
+    let sig_of_type type_decl =
+      let typ = Ppx_deriving.core_type_of_type_decl type_decl in
       let polymorphize = Ppx_deriving.poly_arrow_of_type_decl
-              (fun var -> [%type: [%t var] -> [%t var] -> bool]) type_ in
-      [Sig.value (Val.mk (mknoloc ("equal_"^type_.ptype_name.txt))
+              (fun var -> [%type: [%t var] -> [%t var] -> bool]) type_decl in
+      [Sig.value (Val.mk (mknoloc (Ppx_deriving.mangle_type_decl (`Prefix "equal") type_decl))
                   (polymorphize [%type: [%t typ] -> [%t typ] -> bool]))]
     in
     Ppx_deriving.catch (fun () ->